Chore: Refactor devtools panel top-level hooks
