+++ packages/extension-browser/src/devtools/panel/utils/themes.ts
@@ -1,0 +1,16 @@
+import { browser, document } from '../../../shared/globals';
+
+/** Align active styles with the current devtools theme. */
+export const syncTheme = () => {
+    const onThemeChanged = (theme: string) => {
+        document.documentElement!.setAttribute('data-theme', theme);
+    };
+
+    // Watch for notification of theme changes.
+    if (browser.devtools.panels.onThemeChanged) {
+        browser.devtools.panels.onThemeChanged.addListener(onThemeChanged);
+    }
+
+    // Set the initial theme.
+    onThemeChanged(browser.devtools.panels.themeName);
+};
